Add an odhcp6c script for updating radvd configuration
I have two IPv6 prefixes, one static from a tunnel and another dynamic from ISP,
and I want different M flags for each prefix. This is not an easy task with odhcpd,
while radvd can handle this just fine.

This script acts as an odhcp6c callback script, and should be placed at
/etc/odhcp6c.user. It finds the ISP prefix on LAN interface, and set it in radvd conf.
If upstream DHCPv6 lease gets expired, it will set `ignored` in radvd so that no stale
prefix could be broadcast.

The logic has not been refined - it seems to work, but no way near a perfect (or correct)
implementation.